Basic DMFT script written
This script has been tested to show that it runs, although I am
yet to test if it shows physical results.
It does a DMFT loop for the Hubbard model on the kagome lattice
(or any other DOS you feed into it, but I have a preset for the
kagome lattice). It saves the results to a HDF5 archive. Also saves
simulation metadata (computational parameters, system parameters,
TRIQS and CTHYB version numbers).