Perturbation order histograms. Modularise DMFT.loop() writing
Now have an option to record perturbation order histograms. (faketriqs doesn't read
the Histogram class, but it's irrelevant because the perturbation order histograms
have the x-axis align with the indices anyway so you can just treat it as an array.)
I've also refactored a portion of DMFT.loop(), namely the part where we write the
data from each loop, such that changes in DMFTHubbard automatically propagate to DMFTSubstrate.